Tidied up UI more, reorganised menu, cleaned edit and view pages for users/groups. Added a lock check in groups before modifying (added method to set_lock_status) does not yet do anything

+++ app/controllers/groups_controller.rb
@@ -18,7 +18,7 @@ class GroupsController < ApplicationController
         if @group.save
             #do stuff the devise controller would
             flash[:success] = "New Group!"
-            redirect_to root_url
+            redirect_to @group
         else
             render 'new'
         end
@@ -30,7 +30,10 @@ class GroupsController < ApplicationController
 
     def update
         @group = Group.find(params[:id])
-        if @group.update_attributes(params[:group])
+        if @group.is_locked?
+            flash[:error] = "Group is locked"
+            redirect_to @group
+        elsif @group.update_attributes(params[:group])
             flash[:success] = "Updated!"
             redirect_to @group
         else
@@ -41,11 +44,20 @@ class GroupsController < ApplicationController
     def destroy
         #Does not use @group for some reason described in tutorials
         deleting_group = Group.find(params[:id])
-        #Need error check
-        #Should do group deletion etc...
-        deleting_group.destroy
-    
-        redirect_to groups_path
-        flash[:success] = "Group has been deleted"
+        if deleting_group.is_locked?
+            flash[:error] = "Group is locked"
+            redirect_to deleting_group
+        else
+            #Need error check
+            deleting_group.destroy unless deleting_group.is_locked
+            redirect_to groups_path
+            flash[:success] = "Group has been deleted"
+        end
+    end
+    #is this the right way (I think so)
+    def set_lock_status
+        @group = Group.find(params[:id])
+        @group.toggle_lock
+        redirect_to @group
     end
 end
